--- controllers/contacts/updateStatusContact.js
@@ -1,20 +1,0 @@
-const { Contact } = require("../../models/book");
-
-const { HttpError } = require("../../helpers");
-
-const { ctrlWrapper } = require("../../decorators");
-
-const updateStatusContact = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, `${contactId} not found`);
-  }
-  res.json({ result });
-};
-
-module.exports = {
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-};
