+++ controllers/controllers.js
@@ -1,0 +1,67 @@
+// const contactsService = require("../models/contacts");
+
+const { Contact } = require("../models/book");
+
+const { HttpError } = require("../helpers");
+
+const { ctrlWrapper } = require("../decorators");
+
+const getAllContacts = async (req, res) => {
+  const result = await Contact.find();
+  res.json(result);
+};
+
+const getContactById = async (req, res) => {
+  const { contactId } = req.params;
+
+  const result = await Contact.findById(contactId);
+  if (!result) {
+    throw HttpError(404, `${contactId} not found`);
+  }
+  res.json(result);
+};
+
+const addContact = async (req, res) => {
+  const result = await Contact.create(req.body);
+  res.status(201).json(result);
+};
+
+const deleteContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndRemove(contactId);
+
+  if (!result) {
+    throw HttpError(404, `${contactId} not found`);
+  }
+  res.json({ message: "Delete success" });
+};
+
+const updateContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
+    new: true,
+  });
+  if (!result) {
+    throw HttpError(404, `${contactId} not found`);
+  }
+  res.json({ result });
+};
+const updateStatusContact = async (req, res) => {
+  const { contactId } = req.params;
+  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
+    new: true,
+  });
+  if (!result) {
+    throw HttpError(404, `${contactId} not found`);
+  }
+  res.json({ result });
+};
+
+module.exports = {
+  getAllContacts: ctrlWrapper(getAllContacts),
+  getContactById: ctrlWrapper(getContactById),
+  addContact: ctrlWrapper(addContact),
+  deleteContact: ctrlWrapper(deleteContact),
+  updateContact: ctrlWrapper(updateContact),
+  updateStatusContact: ctrlWrapper(updateStatusContact),
+};
